requirements.txt
APScheduler==3.10.4
bcrypt==4.2.0
blinker==1.9.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzlocal==5.2
Werkzeug==3.1.3
